finished testing TimeSlotService

+++ VehicleRepairShop-Backend/src/main/java/vehiclerepairshop/service/TimeSlotService.java
@@ -27,6 +27,29 @@ public class TimeSlotService {
 		 */
 		@Transactional
 		public TimeSlot createTimeSlot(Time startTime, Time endTime, Date startDate, Date endDate) {
+			String error = "";
+			if (startDate == null) {
+				error = error + "start date cannot be empty! ";
+			}
+			if (startTime == null) {
+				error = error + "start time cannot be empty! ";
+			}
+			if (endTime == null) {
+				error = error + "end time cannot be empty! ";
+			}
+			if (endDate == null) {
+				error = error + "end date cannot be empty! ";
+			}
+			if (endTime != null && startTime != null && startTime.after(endTime)) {
+				error = error + "start time can't be after end time";
+			}
+			if (endDate != null && startDate != null &&startDate.after(endDate)){
+				error = error + "start date can't be after end date";
+			}
+			error = error.trim();
+		    if (error.length() > 0) {
+		        throw new IllegalArgumentException(error);
+		    }
 			TimeSlot timeSlot = new TimeSlot();
 			timeSlot.setEndDate(endDate);
 			timeSlot.setEndTime(endTime);
@@ -49,22 +72,90 @@ public class TimeSlotService {
 		}
 		
 		/**
+		 * @author chengchen
+		 */
+		@Transactional
+		public TimeSlot getTimeSlot(int id) {
+			String error = "";
+			if (timeslotRepository.findByTimeSlotId(id)==null) {
+				error = error + "timeslot not found";
+			}
+			error = error.trim();
+		    if (error.length() > 0) {
+		        throw new IllegalArgumentException(error);
+		    }
+			return timeslotRepository.findByTimeSlotId(id);
+		}
+		
+		/**
 		 * 
 		 * @param timeslotId
 		 * @author chengchen
 		 */
 		@Transactional
-		public void deleteTimeSlot(int timeslotId) {
+		public TimeSlot deleteTimeSlot(int timeslotId) { 
+			String error = "";
+			if (timeslotRepository.findByTimeSlotId(timeslotId)==null) {
+				error = error + "timeslot not found";
+			}
+			error = error.trim();
+		    if (error.length() > 0) {
+		        throw new IllegalArgumentException(error);
+		    }
 			TimeSlot timeSlot = timeslotRepository.findByTimeSlotId(timeslotId);
 			timeslotRepository.delete(timeSlot);
+			return timeSlot;
 		}
 		
 		/**
 		 * @author chengchen
+		 * @return
 		 */
 		@Transactional
-		public void deleteAllTimeSlots() {
+		public List<TimeSlot> deleteAllTimeSlot(){
+			Iterable<TimeSlot> timeSlots = timeslotRepository.findAll();
 			timeslotRepository.deleteAll();
+			return toList(timeSlots);
+		}
+		
+		@Transactional
+		public TimeSlot updateTimeSlot(int id, Time startTime, Time endTime, Date startDate, Date endDate) {
+			String error = "";
+			TimeSlot timeSlot = timeslotRepository.findByTimeSlotId(id);
+			if (startDate == null) {
+				error = error + "start date cannot be empty! ";
+			}
+			if (startTime == null) {
+				error = error + "start time cannot be empty! ";
+			}
+			if (endTime == null) {
+				error = error + "end time cannot be empty! ";
+			}
+			if (endDate == null) {
+				error = error + "end date cannot be empty! ";
+			}
+			if (endTime != null && startTime != null && startTime.after(endTime)) {
+				error = error + "start time can't be after end time";
+			}
+			if (endDate != null && startDate != null &&startDate.after(endDate)){
+				error = error + "start date can't be after end date";
+			}
+			if (timeSlot == null) {
+				error = error + "non-existing time slot";
+			}
+			error = error.trim();
+		    if (error.length() > 0) {
+		        throw new IllegalArgumentException(error);
+		    }
+
+			timeSlot.setEndDate(endDate);
+			timeSlot.setEndTime(endTime);
+			timeSlot.setStartDate(startDate);
+			timeSlot.setStartTime(startTime);
+			timeSlot.setTimeSlotId(id);
+			timeslotRepository.save(timeSlot);
+			
+			return timeSlot;
 		}
 		
 		// helper method that converts iterable to list
